refactor(audio): split player module and add loop progress tests
